requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
langgraph>=0.2.16
langchain>=0.2.16
langchain-openai>=0.1.25
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1
httpx>=0.25.2
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
pyyaml>=6.0.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
PyPDF2>=3.0.1
python-docx>=0.8.11
pdfplumber>=0.9.0